python/pyproject.toml
[project]
name = "mcp-streamable-http-server"
version = "0.1.0"
description = "HTTP Streamable based example MCP Server"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }

# Application dependencies
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "azure-storage-blob",
    "python-dotenv",
    "python-pptx",
    "openpyxl",
    "requests"
]

[tool.uv]
# Only using uv for package management
